requirements.txt
numpy
pandas
opencv-python
pillow
pyyaml
matplotlib
jupyter
ipykernel


ultralytics